requirements.txt
pandas==1.3.5
scikit-learn==1.2.1
pyYAML==6.0.0
boto3==1.24.28
pyspark==3.3.1
pyarrow==14.0.1
